fix(runtime): transactional guarded step, explicit fp32 clock, extended cache proof
- guarded_optimizer_step is now fully transactional: pre-step rejection
  validates loss/clip/params/grads/norm before any mutation; parameters
  and a deep exact optimizer-state snapshot are restored bit-exactly
  when the step raises or post-step params/state go non-finite
- LocalizedRecurrence clock trainables are explicitly fp32 regardless of
  global default dtype (bf16-default regression test added)
- tiny-Qwen cached-vs-reference equivalence proven for full AND genuinely
  localized intervals, before and after a nonzero fp32 bundle roundtrip,
  with cache-storage/position consumption sabotage assertions
- latent_run: import BestCheckpointTracker/guarded_optimizer_step (were
  referenced without import since 5559f32)

+++ latent_lab/backends/localized.py
@@ -215,7 +215,9 @@ class LocalizedRecurrence:
         self.guard = VocabGuard(model, tokenizer)
 
         dev = next(model.parameters()).device
-        self.clock = torch.nn.Embedding(max_k + 1, hidden, device=dev)
+        # clock trainables are explicitly fp32; never the global default dtype
+        self.clock = torch.nn.Embedding(max_k + 1, hidden, device=dev,
+                                        dtype=torch.float32)
         torch.nn.init.zeros_(self.clock.weight)
 
         self.injected = inject_lora(
